+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/client/MantaClient.java
@@ -196,7 +196,7 @@ public class MantaClient implements AutoCloseable {
      *
      * Users opting into advanced configuration (i.e. not passing {@code null} as the second parameter)
      * should be comfortable with the internals of {@link CloseableHttpClient} and accept that we can only make a
-     * best effort to support all possible use-cases. For example, uses may pass in a builder which is wired to a
+     * best effort to support all possible use-cases. For example, users may pass in a builder which is wired to a
      * {@link org.apache.http.impl.conn.BasicHttpClientConnectionManager} and effectively make the client
      * single-threaded by eliminating the connection pool. Bug or feature? You decide!
      *
@@ -214,13 +214,13 @@ public class MantaClient implements AutoCloseable {
      *
      * Users opting into advanced configuration (i.e. not passing {@code null} as the second parameter)
      * should be comfortable with the internals of {@link CloseableHttpClient} and accept that we can only make a
-     * best effort to support all possible use-cases. For example, uses may pass in a builder which is wired to a
+     * best effort to support all possible use-cases. For example, users may pass in a builder which is wired to a
      * {@link org.apache.http.impl.conn.BasicHttpClientConnectionManager} and effectively make the client
      * single-threaded by eliminating the connection pool. Bug or feature? You decide!
      *
      * @param config The configuration context that provides all of the configuration values
      * @param connectionFactoryConfigurator pre-configured objects for use with a MantaConnectionFactory (or null)
-     * @param httpHelper helper object for executing http requests (or null)
+     * @param httpHelper helper object for executing http requests (or null to build one ourselves)
      */
     MantaClient(final ConfigContext config,
                 final MantaConnectionFactoryConfigurator connectionFactoryConfigurator,
@@ -268,7 +268,7 @@ public class MantaClient implements AutoCloseable {
 
         if (httpHelper != null) {
             this.httpHelper = httpHelper;
-        } else if (BooleanUtils.isTrue(config.isClientEncryptionEnabled())) {
+        } else if (BooleanUtils.isTrue(this.config.isClientEncryptionEnabled())) {
             this.httpHelper = new EncryptionHttpHelper(connectionContext, requestFactory, config);
         } else {
             this.httpHelper = new StandardHttpHelper(
@@ -343,15 +343,28 @@ public class MantaClient implements AutoCloseable {
      *
      * @param rawPath The fully qualified path of the Manta object.
      * @throws IOException If an IO exception has occurred.
-     * @throws MantaClientHttpResponseException                If an HTTP status code {@literal > 300} is returned.
+     * @throws MantaClientHttpResponseException If a HTTP status code other than {@code 200 | 202 | 204} is encountered
      */
     public void delete(final String rawPath) throws IOException {
+        this.delete(rawPath, null);
+    }
+
+    /**
+     * Deletes an object from Manta, with optional headers.
+     *
+     * @param rawPath The fully qualified path of the Manta object.
+     * @param requestHeaders HTTP headers to attach to request (may be null)
+     * @throws IOException If an IO exception has occurred.
+     * @throws MantaClientHttpResponseException If a HTTP status code other than {@code 200 | 202 | 204} is encountered
+     */
+    public void delete(final String rawPath,
+                       final MantaHttpHeaders requestHeaders) throws IOException {
         Validate.notBlank(rawPath, "rawPath must not be blank");
 
         String path = formatPath(rawPath);
         LOG.debug("DELETE {}", path);
 
-        httpHelper.httpDelete(path);
+        httpHelper.httpDelete(path, requestHeaders);
     }
 
     /**