+++ java-manta-client-unshaded/src/test/java/com/joyent/manta/http/StandardHttpHelperTest.java
@@ -26,8 +26,10 @@ import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.message.BasicHeader;
 import org.bouncycastle.crypto.Digest;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.testng.Assert;
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
@@ -36,11 +38,18 @@ import java.util.Base64;
 
 import static com.joyent.manta.client.MantaObjectResponse.DIRECTORY_RESPONSE_CONTENT_TYPE;
 import static com.joyent.manta.util.UnitTestConstants.UNIT_TEST_URL;
+import static org.apache.http.HttpHeaders.IF_MATCH;
+import static org.apache.http.HttpStatus.SC_BAD_REQUEST;
+import static org.apache.http.HttpStatus.SC_CREATED;
+import static org.apache.http.HttpStatus.SC_NO_CONTENT;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.reset;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+@Test
 public class StandardHttpHelperTest {
 
     @Mock
@@ -58,6 +67,7 @@ public class StandardHttpHelperTest {
     public void setup() throws Exception {
         MockitoAnnotations.initMocks(this);
         config = new StandardConfigContext().setMantaURL(UNIT_TEST_URL);
+        reset(client, response, connCtx, statusLine);
 
         when(connCtx.getHttpClient())
                 .thenReturn(client);
@@ -72,7 +82,11 @@ public class StandardHttpHelperTest {
                 .thenReturn(new Header[]{});
     }
 
-    @Test
+    @AfterMethod
+    public void teardown() throws Exception {
+        Mockito.validateMockitoUsage();
+    }
+
     public void testHttpPutValidatesResponseCodeSuccessfully() throws Exception {
         when(statusLine.getStatusCode())
                 .thenReturn(HttpStatus.SC_NO_CONTENT);
@@ -89,7 +103,6 @@ public class StandardHttpHelperTest {
         Assert.assertNotNull(put);
     }
 
-    @Test
     public void testHttpPutValidatesResponseCodeAndThrowsWhenInvalid() throws Exception {
         when(statusLine.getStatusCode())
                 .thenReturn(HttpStatus.SC_OK);
@@ -103,7 +116,6 @@ public class StandardHttpHelperTest {
                 helper.httpPut("/path", null, NoContentEntity.INSTANCE, null));
     }
 
-    @Test
     public void testHttpPutChecksumsSuccessfully() throws Exception {
         when(statusLine.getStatusCode())
                 .thenReturn(HttpStatus.SC_NO_CONTENT);
@@ -144,7 +156,6 @@ public class StandardHttpHelperTest {
         Assert.assertNotNull(put);
     }
 
-    @Test
     public void testHttpPutChecksumsCompareDifferentlyFails() throws Exception {
         when(statusLine.getStatusCode())
                 .thenReturn(HttpStatus.SC_NO_CONTENT);
@@ -179,7 +190,6 @@ public class StandardHttpHelperTest {
                 helper.httpPut("/path", null, new ByteArrayEntity(contentBytes), null));
     }
 
-    @Test
     public void testHttpPutThrowsWhenChecksumRequestedButNotReturned() throws Exception {
         when(statusLine.getStatusCode())
                 .thenReturn(HttpStatus.SC_NO_CONTENT);
@@ -196,7 +206,6 @@ public class StandardHttpHelperTest {
                 helper.httpPut("/path", null, NoContentEntity.INSTANCE, null));
     }
 
-    @Test
     public void throwsAppropriateExceptionWhenStreamingObjectThatIsDir() {
         final String path = "/user/stor/a-dir";
         final StandardHttpHelper helper = new StandardHttpHelper(connCtx, config);
@@ -222,5 +231,41 @@ public class StandardHttpHelperTest {
 
         Assert.assertThrows(MantaUnexpectedObjectTypeException.class, () ->
                 helper.httpRequestAsInputStream(get, headers));
+    }
+
+    @Test(expectedExceptions = MantaClientHttpResponseException.class,
+          expectedExceptionsMessageRegExp = ".*code.*expected.*got.*" + SC_CREATED + ".*")
+    public void deleteValidatesSuccessResponseCode() throws Exception {
+        when(statusLine.getStatusCode())
+                .thenReturn(SC_CREATED); // this is a nonsensical response code for DELETE
+
+        final StandardHttpHelper helper = new StandardHttpHelper(connCtx, config);
+
+        helper.httpDelete("/path");
+    }
+
+    @Test(expectedExceptions = MantaClientHttpResponseException.class)
+    public void deleteValidatesErrorResponseCode() throws Exception {
+        when(statusLine.getStatusCode())
+                .thenReturn(SC_BAD_REQUEST);
+
+        final StandardHttpHelper helper = new StandardHttpHelper(connCtx, config);
+
+        helper.httpDelete("/path");
+    }
+
+
+    public void deleteWithHeadersPassesAlongHeaders() throws Exception {
+        when(statusLine.getStatusCode())
+                .thenReturn(SC_NO_CONTENT);
+
+        final StandardHttpHelper helper = new StandardHttpHelper(connCtx, config);
+        final MantaHttpHeaders ifMatchHeader = new MantaHttpHeaders();
+        final String etag = "foo";
+        ifMatchHeader.setIfMatch(etag);
+
+        helper.httpDelete("/path", ifMatchHeader);
+
+        verify(client).execute(argThat(r -> etag.equals(r.getFirstHeader(IF_MATCH).getValue())));
     }
 }