+++ java-manta-it/src/test/java/com/joyent/manta/http/MantaHttpHeadersIT.java
@@ -9,6 +9,7 @@ package com.joyent.manta.http;
 
 import com.joyent.manta.client.MantaClient;
 import com.joyent.manta.client.MantaObject;
+import com.joyent.manta.client.MantaObjectResponse;
 import com.joyent.manta.config.ConfigContext;
 import com.joyent.manta.config.IntegrationTestConfigContext;
 import com.joyent.manta.exception.MantaClientHttpResponseException;
@@ -16,6 +17,8 @@ import com.joyent.test.util.MantaAssert;
 import com.joyent.test.util.MantaFunction;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.ObjectUtils;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.http.HttpStatus;
 import org.testng.Assert;
 import org.testng.SkipException;
 import org.testng.annotations.AfterClass;
@@ -29,6 +32,10 @@ import java.util.UUID;
 
 import static com.joyent.manta.exception.MantaErrorCode.INVALID_ROLE_TAG_ERROR;
 import static com.joyent.manta.util.MantaUtils.asString;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertTrue;
+import static org.testng.Assert.expectThrows;
 
 /**
  * Integration tests for verifying the behavior of HTTP header assignment.
@@ -41,6 +48,7 @@ import static com.joyent.manta.util.MantaUtils.asString;
 @SuppressWarnings("Duplicates")
 @Test(groups = { "headers" })
 public class MantaHttpHeadersIT {
+
     private static final String TEST_DATA = "EPISODEII_IS_BEST_EPISODE";
 
     private MantaClient mantaClient;
@@ -73,27 +81,25 @@ public class MantaHttpHeadersIT {
         IntegrationTestConfigContext.cleanupTestDirectory(mantaClient, testPathPrefix);
     }
 
-    @Test
     public void cantSetUnknownTags() throws IOException {
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         final Set<String> roles = new HashSet<>();
         roles.add("thistestprobablydoesntexist");
         headers.setRoles(roles);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         MantaAssert.assertResponseFailureStatusCode(409, INVALID_ROLE_TAG_ERROR,
                 (MantaFunction<Object>) () -> mantaClient.put(path, TEST_DATA, headers));
     }
 
-    @Test
     public void canSetSingleRoleTag() throws IOException {
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         final Set<String> roles = new HashSet<>();
         roles.add(primaryRoleName);
         headers.setRoles(roles);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         try {
             mantaClient.put(path, TEST_DATA, headers);
@@ -122,7 +128,6 @@ public class MantaHttpHeadersIT {
         }
     }
 
-    @Test
     public void canSetMultipleRoleTags() throws IOException {
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         final Set<String> roles = new HashSet<>();
@@ -130,7 +135,7 @@ public class MantaHttpHeadersIT {
         roles.add(secondaryRoleName);
         headers.setRoles(roles);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         try {
             mantaClient.put(path, TEST_DATA, headers);
@@ -160,7 +165,6 @@ public class MantaHttpHeadersIT {
         }
     }
 
-    @Test
     public void canOverwriteRoleTags() throws IOException {
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         final Set<String> roles = new HashSet<>();
@@ -168,7 +172,7 @@ public class MantaHttpHeadersIT {
         roles.add(secondaryRoleName);
         headers.setRoles(roles);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         try {
             mantaClient.put(path, TEST_DATA, headers);
@@ -215,7 +219,6 @@ public class MantaHttpHeadersIT {
         }
     }
 
-    @Test
     public void canClearRoles() throws IOException {
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         final Set<String> roles = new HashSet<>();
@@ -223,7 +226,7 @@ public class MantaHttpHeadersIT {
         roles.add(secondaryRoleName);
         headers.setRoles(roles);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         try {
             mantaClient.put(path, TEST_DATA, headers);
@@ -255,13 +258,12 @@ public class MantaHttpHeadersIT {
         }
     }
 
-    @Test
     public void canSetDurability() throws IOException {
         final int durability = 3;
         final MantaHttpHeaders headers = new MantaHttpHeaders();
         headers.setDurabilityLevel(durability);
 
-        String path = String.format("%s/%s", testPathPrefix, UUID.randomUUID());
+        String path = generatePath();
 
         mantaClient.put(path, TEST_DATA, headers);
 
@@ -269,5 +271,37 @@ public class MantaHttpHeadersIT {
         MantaHttpHeaders actualHeaders = object.getHttpHeaders();
         Assert.assertEquals(actualHeaders.getDurabilityLevel().intValue(), durability,
                 "Durability not set to value on put");
+    }
+
+    public void canFailToDeleteDirectoryOnBadIfMatch() throws IOException {
+        final String path = generatePath();
+
+        final MantaObjectResponse empty = mantaClient.put(path, new byte[0]);
+        // the etag reversed should not be equal to the etag (so we can fail if-match on purpose)
+        assertFalse(empty.getEtag().equals(StringUtils.reverse(empty.getEtag())));
+
+        final MantaHttpHeaders headers = new MantaHttpHeaders();
+
+        // fail on bad if-match
+        headers.setIfMatch(StringUtils.reverse(empty.getEtag()));
+        final MantaClientHttpResponseException badIfMatchEx = expectThrows(
+                MantaClientHttpResponseException.class,
+                () -> mantaClient.delete(path, headers));
+
+        assertEquals(badIfMatchEx.getStatusCode(), HttpStatus.SC_PRECONDITION_FAILED);
+
+        // the object should still exist
+        assertTrue(mantaClient.existsAndIsAccessible(path));
+
+        // set the correct If-Match header
+        headers.setIfMatch(empty.getEtag());
+        mantaClient.delete(path, headers);
+
+        // the object should not exist
+        assertFalse(mantaClient.existsAndIsAccessible(path));
+    }
+
+    private String generatePath() {
+        return String.format("%s/%s", testPathPrefix, UUID.randomUUID());
     }
 }